short int

+++ map.h
@@ -6,13 +6,13 @@
 typedef struct _map map;
 
 map* readMap(FILE *);
-int varA (map *, int, int);
-void compA (map *, int, int, int *);
-int inMapCheck (map *, int, int);
+short int varA (map *, short int, short int);
+void compA (map *, short int, short int, short int *);
+short int inMapCheck (map *, short int, short int);
 char getMode(map *);
-int getPOI(map *, int, int);
+short int getPOI(map *, short int, short int);
 void modeVarA(map *, FILE *);
-int validMove(map *, int);
+short int validMove(map *, short int);
 void modeVarB(map *, FILE *);
 void freeMap(map *);
 void printerror(map *,FILE *);
